requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
psycopg2-binary>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.15.0
streamlit-folium>=0.15.0
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
neo4j>=5.15.0
flake8>=6.1.0
testcontainers[postgres]>=4.0.0
testcontainers[neo4j]>=4.0.0

